Implement leader key
I did not use QMK's leader implementation, because it's completely
absurd. It works on a timeout system where you have to type the entire
sequence within the configured time. Make it too short, and you'll
struggle to activate longer sequences. Make it too long, and you'll have
to wait for the timeout before anything happens. Since the sequence has
to be triggered by a dedicated leader key, imposing a time limit is
pointless anyways.

This version has the following advantages:

* No time limit
* Leader sequence can be canceled with a second tap of the leader key
* A valid sequence triggers the action immediately
* An invalid sequence terminates the sequence (only when max length is
  exceeded---not ideal)
* Provides audio feedback when a sequence is canceled
* Cleaner DSL for defining sequences/actions

Of course, with this system you can't have a terminable sequence that is
a prefix of another sequence without making the longer sequence
unreachable, which is the one "advantage" of QMK's implementation, if
you can call it that. A more complex implementation could take a hybrid
approach that would emulate Vim's behavior when encountering an
ambiguous sequence.